* Version_1.1: (43 commits)
  Update - Build 5
  Revise - Unread article flag
  Update - AnswerDetail: headerImageView <--> contentView
  Update
  Update - Fade in
  Revise - 3 bottom buttons
  Update - Hide statusBar
  Update - Build 4
  Update - Build 3
  Update
  Update - QQ platform login
  Version - 1.1
  Update
  Remove - timeLabel hidden = YES
  Remove - Date label
  Replace - trailer viedo
  Add - Happy mascot
  Update - Rank label
  Update - Mascot unread flag
  Update - load timing
  ...

# Conflicts:
#	NineZeroProject.xcodeproj/project.pbxproj
#	NineZeroProject/Model/HTModel.h
#	Pods/Pods.xcodeproj/xcuserdata/SinLemon.xcuserdatad/xcschemes/xcschememanagement.plist

+++ NineZeroProject/Service/SKService/SKQuestionService.m
@@ -113,8 +113,8 @@
                                  @"action": [SKCGIManager question_getQuestionList_Successful_Action],
                                  @"user_id": [[SKStorageManager sharedInstance] getUserID],
                                  @"login_token": [[SKStorageManager sharedInstance] getUserToken],
-                                 @"page": [NSString stringWithFormat:@"%lu", page],
-                                 @"page_size": [NSString stringWithFormat:@"%lu", count]
+                                 @"page": [NSString stringWithFormat:@"%lu", (unsigned long)page],
+                                 @"page_size": [NSString stringWithFormat:@"%lu", (unsigned long)count]
                                  };
     
     NSMutableURLRequest* request = [[AFJSONRequestSerializer serializer] requestWithMethod:@"POST" URLString:@"http://90app-test.daoapp.io/api/question" parameters:bodyObject error:NULL];
